+++ api/microchip/pic16f6xx_18p/pin_fn/pin_fn.c.v
@@ -24,11 +24,11 @@ module pin_fn
 @[as_macro] pub const b7 = 15
 
 // IRQ (INT) pin
-@[as_macro] pub const irq = 8
+@[as_macro] pub const irq = u8(8)
 
 // pin modes as macros
-@[as_macro] pub const output = 0
-@[as_macro] pub const input = 1
+@[as_macro] pub const output = u8(0)
+@[as_macro] pub const input = u8(1)
 
 // port names as macros
 @[as_macro] pub const port_a = C.A
