+++ api/microchip/pic16f8x/pin_fast/pin_fast.c.v
@@ -21,10 +21,10 @@ module pin_fast
 @[as_macro] pub const b7 = 15
 
 // IRQ (INT) pin
-@[as_macro] pub const irq = 8
+@[as_macro] pub const irq = u8(8)
 
-@[as_macro] pub const output = 0    // pin modes
-@[as_macro] pub const input = 1
+@[as_macro] pub const output = u8(0)    // pin modes
+@[as_macro] pub const input = u8(1)
 
 @[as_macro] pub const port_a = C.A  // port ids
 @[as_macro] pub const port_b = C.B
