+++ api/weact_studio/esp32_c3fh4/pin/pin.c.v
@@ -7,19 +7,19 @@
 // Pin management functions Arduino devices
 module pin
 
-@[as_macro] pub const d0 = 0
-@[as_macro] pub const d1 = 1
-@[as_macro] pub const d2 = 2
-@[as_macro] pub const d3 = 3
-@[as_macro] pub const d4 = 4
-@[as_macro] pub const d5 = 5
-@[as_macro] pub const d6 = 6
-@[as_macro] pub const d7 = 7
-@[as_macro] pub const d8 = 8
-@[as_macro] pub const d9 = 9
-@[as_macro] pub const d10 = 10
+@[as_macro] pub const d0 = u8(0)
+@[as_macro] pub const d1 = u8(1)
+@[as_macro] pub const d2 = u8(2)
+@[as_macro] pub const d3 = u8(3)
+@[as_macro] pub const d4 = u8(4)
+@[as_macro] pub const d5 = u8(5)
+@[as_macro] pub const d6 = u8(6)
+@[as_macro] pub const d7 = u8(7)
+@[as_macro] pub const d8 = u8(8)
+@[as_macro] pub const d9 = u8(9)
+@[as_macro] pub const d10 = u8(10)
 
-@[as_macro] pub const d18 = 18
-@[as_macro] pub const d19 = 19
-@[as_macro] pub const d20 = 20
-@[as_macro] pub const d21 = 21
+@[as_macro] pub const d18 = u8(18)
+@[as_macro] pub const d19 = u8(19)
+@[as_macro] pub const d20 = u8(20)
+@[as_macro] pub const d21 = u8(21)
